Update to Dash 2.5.1 from Dash Labs multipage
Now that Dash 2.5 has come out and Multipage apps are supported
natively in the core library, this commit switches from the in-development
dashlabs_multipage_app interface to the Dash version.

Along with this change, we are able to remove previous annoyance code
such as having both an index.py and an app.py file- now everything
can be handled in the app.py file and index.py has been removed.

Because of how python files are preprocessed, validation_layout has been
removed in this commit- whether it is necessary or not in the future will
be up to whether this leads to unforseen problems.

+
Removed .DS_Store
Should have been in the .gitignore from the start.

Signed-off-by: JamesKunstle <jkunstle@bu.edu>