Add pytest tests, GH workflow to test dev, promote dev to main
Two dummy tests were added and examples of pytest fixtures are included.

Workflow script is included to build the container that we want to test
our code in for our 'dev' branch, on success this gets pushed to :dev from
:test.

Also adds workflow to promote our image from :dev to :main on quay
so that we automatically deploy our most updated image when we merge our
codebase to main.

Signed-off-by: JamesKunstle <jkunstle@bu.edu>