update: smaller commits query, optimized commit callback
The query that downloaded commit data was slow because it
had to download a lot of currently unused data.

Removed unused columns and added more efficient compression
to POSIX timestamps when caching data.

Updated commits_figure callback to be more compartmentalized
and to use a more efficient representation of the same data
with a bar-plot rather than a histogram.

Signed-off-by: James Kunstle <jkunstle@bu.edu>