remove mypy from linting, clean up files.
mypy was returning error: "8Knot is not a valid Python package name."
This is true, but we want to ignore this error.
However, we can't ignore it because mypy won't report the error message
that the error corresponds to, so we can't explicitly ignore that error.

Therefore, we're removing mypy from pre-commit linting for the
forseeable future. Other linting w/ black and pre-commit will
still happen.

Signed-off-by: James Kunstle <jkunstle@bu.edu>