Deactivate home page metrics page;
Promote Info page to index-route, e.g. '/'

If applied, we won't register the 'home' page with the Dash application.
This will remove it from being served.
We also promote the 'info' page the the 'index' route so that it's what
users will see when they first go to 8Knot.

Rationale for this decision is in the file docstring of:
pages/home/home.py

Will also be copied in PR that represents this commit.

Signed-off-by: James Kunstle <jkunstle@bu.edu>