removes use of pconfig, purely environment
fixes #253
previously, pconfig was an object passed to celery workers that
contained database access credentials in lieu of sending a global-scope
instantiated AugurManager object.

this is replaced by created a new AugurManager object in each Celery
worker process new from the available environment variables.

This should be more secure because we're not passing access parameters,
in plaintext, between processes, as worker parameters.

Signed-off-by: James Kunstle <jkunstle@bu.edu>