fix queries to accept null timestamps
queries check to make sure event timestamps didn't happen in the future.
this patch adds a check to accept a rows with NULL in timestamp
fields so that if an event hasn't completed (issue or PR closed)
the row is still accepted.

also updates the cache tables to use consistent data types.

Signed-off-by: James Kunstle <jkunstle@bu.edu>